requirements.txt
Flask[async]==2.3.2
flask-socketio==5.3.4
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
transformers==4.29.2
indic-nlp-library==0.91
aiohttp==3.8.4
python-dotenv==1.0.0
numpy==1.24.3
accelerate==0.19.0
